fix diagonal movement rate
Note that `scale_to` gracefully handles the zero vector, so no special
case is needed.

Jamie and I discussed and both had the gut feeling that leaving the
`direction` property unchanged from now (so that its elements are still
always +1, -1, or 0) was easier for implementation but could also make
sense when explained.

Closes: #709